# api/openapi.yaml
openapi: 3.1.0
info:
  title: sing-box WebUI Local API
  version: 0.1.0
  description: Local loopback API for the sing-box WebUI control plane.
servers:
  - url: http://127.0.0.1:11872
paths:
...
  /api/v1/links/clear:
    post:
      operationId: clearLinks
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '200':
          description: Connection cache cleared. Monitoring continues.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [status]
                properties:
                  status:
                    type: string
        default:
          $ref: '#/components/responses/Error'
...
components:
  parameters:
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, requestId]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
        requestId:
          type: string